Move out utility and previously-mutating methods into TxSetUtils
